@@ -5,24 +5,24 @@
   "homepage": "https://coderberg.github.io/web-developer-tools",
   "dependencies": {
     "@reduxjs/toolkit": "^1.7.1",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^13.5.0",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^14.2.0",
     "@types/jest": "^24.9.1",
-    "@types/node": "^12.20.37",
-    "@types/react": "^16.14.21",
-    "@types/react-dom": "^16.9.14",
+    "@types/node": "^17.0.36",
+    "@types/react": "^18.0.9",
+    "@types/react-dom": "^18.0.5",
     "@types/react-redux": "^7.1.20",
     "@types/react-transition-group": "^4.4.4",
     "generate-password-browser": "^1.1.0",
     "primeflex": "2.0.0",
     "primeicons": "^5.0.0",
-    "primereact": "^7.1.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "primereact": "^8.1.1",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-redux": "^7.2.6",
     "react-router-dom": "^6.2.1",
-    "react-scripts": "5.0.0",
+    "react-scripts": "5.0.1",
     "react-transition-group": "^4.4.2",
     "redux": "^4.1.2",
     "typescript": "^4.1.6"
