@@ -1,19 +1,18 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import './index.css';
 import {Provider} from "react-redux";
+import {createRoot} from "react-dom/client";
 import App from './App';
 import * as serviceWorker from './serviceWorker';
 import store from "./store/configureStore";
 
-ReactDOM.render(
-    <React.StrictMode>
-        <Provider store={store}>
-            <App/>
-        </Provider>
-    </React.StrictMode>,
-    document.getElementById('root')
-);
+const container = document.getElementById('root');
+const root = createRoot(container!);
+root.render(<React.StrictMode>
+    <Provider store={store}>
+        <App/>
+    </Provider>
+</React.StrictMode>);
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
